+++ Sources/PulseUI/Features/Search/Views/ConsoleSearchContextMenu.swift
@@ -19,7 +19,7 @@ struct ConsoleSearchContextMenu: View {
         } label: {
             Image(systemName: "ellipsis.circle")
                 .font(.system(size: 20))
-                .foregroundColor(.blue)
+                .foregroundColor(.accentColor)
         }
     }
 }